+++ Ideals Test Project/Tests/BaseTest.cs
@@ -2,6 +2,7 @@
 using OpenQA.Selenium;
 using OpenQA.Selenium.Chrome;
 using OpenQA.Selenium.Firefox;
+using System.Reflection;
 
 namespace Ideals_Test_Project.Tests
 {
@@ -32,6 +33,24 @@ namespace Ideals_Test_Project.Tests
         public void OneTimeTearDown()
         {
             ReporterHelper.CloseReporter();
+        }
+
+        protected void BaseTearDown()
+        {
+            if (TestContext.CurrentContext.Result.Outcome.Status == NUnit.Framework.Interfaces.TestStatus.Failed)
+            {
+                ReporterHelper.Log(AventStack.ExtentReports.Status.Error, TestContext.CurrentContext.Result.Message);
+
+                Screenshot screenshot = (Driver as ITakesScreenshot).GetScreenshot();
+                string title = TestContext.CurrentContext.Test.Name;
+                string runname = $"{title} {DateTime.Now.ToString("yyyy - MM - dd - HH_mm_ss")}";
+                string screenshotfilename = $"{Assembly.GetEntryAssembly().Location}{runname}.jpg";
+                screenshot.SaveAsFile(screenshotfilename, ScreenshotImageFormat.Jpeg);
+
+                ReporterHelper.SaveScreenshot(screenshotfilename);
+            }
+
+            Driver.Quit();
         }
     }
 }
